+++ src/components/news/CommentDetail.vue
@@ -125,8 +125,8 @@
   import {Indicator} from 'mint-ui';
   import {Header} from 'mint-ui';
 
-  import URL_PARAMS from '../urls-config';
-  import PARAMS from '../../config/index';
+  import URL_PARAMS from '../../urls-config';
+  import PARAMS from '../../../config/index';
 
   export default {
     name: 'comment',
