+++ src/components/news/Search.vue
@@ -145,8 +145,8 @@
 <script>
   import {Indicator} from 'mint-ui';
   import {Header} from 'mint-ui';
-  import { URL as URL_PARAMS } from '../urls-config';
-  import PARAMS from '../../config/index';
+  import { URL as URL_PARAMS } from '../../urls-config';
+  import PARAMS from '../../../config/index';
 
   export default {
     name: 'comment',
