+++ src/components/test/Detail.vue
